+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Calculate Return Statistics
+- Quandl Handyman functions
 ### Fixed
 ### Changed
 - run scaling in HRP Optimizer